+++ src/types/Todo.ts
@@ -3,3 +3,9 @@ export interface Todo {
   title: string;
   completed: boolean;
 }
+
+export enum Filter {
+  all = 'all',
+  active = 'active',
+  completed = 'completed',
+}
